Confirmation accout for adding books

+++ bookreview/views/book.py
@@ -10,14 +10,15 @@ from bookreview.models import Book, Review, User, Comment, Permissions
 book = Blueprint('book', __name__)
 
 
-# @book.before_app_request
-# def check_permissions_post():
-#     if request.method == 'POST' and current_user.is_anonymous:
-#         flash('Войдите в аккаунт', category='warning')
-#         return redirect(url_for('authorization.login'))
-#     if request.method == "POST" and not current_user.confirmed:
-#         flash('Сперва подтвердите аккаунт', category='danger')
-#         return request.args
+def confirm_required(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        if not current_user.confirmed:
+            flash('Подтвердите свой аккаунт', category='warning')
+            return redirect(request.referrer)
+        return func(*args, **kwargs)
+
+    return wrapper
 
 
 @book.route('/books/<int:user_id>', methods=["POST", "GET"])
@@ -31,14 +32,17 @@ def add_book(user_id):
     page = request.args.get('page', 1, type=int)
     books = user.books.paginate(per_page=9, page=page)
     if add_book_form.validate_on_submit():
-        new_book = Book(user_id=current_user.id,
-                        title=add_book_form.title.data,
-                        author=add_book_form.author.data,
-                        cover=bookcover.save(add_book_form.cover.data) if add_book_form.cover.data else None,
-                        description=add_book_form.description.data)
-        db.session.add(new_book)
-        db.session.commit()
-        flash("Книга добавлена", category="success")
+        if current_user.can(Permissions.WRITE):
+            new_book = Book(user_id=current_user.id,
+                            title=add_book_form.title.data,
+                            author=add_book_form.author.data,
+                            cover=bookcover.save(add_book_form.cover.data) if add_book_form.cover.data else None,
+                            description=add_book_form.description.data)
+            db.session.add(new_book)
+            db.session.commit()
+            flash("Книга добавлена", category="success")
+        else:
+            flash('Подтвердите аккаунт, чтоб добавлять книги', category='warning')
         return redirect(url_for('book.add_book', user_id=user_id))
 
     return render_template('add_book.html', form=add_book_form, books=books, user=user)
@@ -48,7 +52,7 @@ def add_book(user_id):
 @login_required
 def delete_book(book_id):
     book_ = Book.query.get_or_404(int(book_id))
-    if current_user.id == book_.user.id:
+    if current_user.id == book_.user.id and current_user.can(Permissions.DELETE):
         db.session.delete(book_)
         db.session.commit()
         flash("Книга удалена", category="warning")
@@ -93,7 +97,7 @@ def write_review(user_id):
     reviews = user.reviews.order_by(Review.date.desc(), Review.popularity.desc()).paginate(per_page=15, page=page)
 
     write_review_form = WriteReview()
-    if write_review_form.validate_on_submit():
+    if write_review_form.validate_on_submit() and current_user.can(Permissions.WRITE):
         review_ = Review(author_id=current_user.id,
                          book_id=write_review_form.select_book.data.id,
                          text=write_review_form.text.data)
@@ -107,9 +111,11 @@ def write_review(user_id):
 
 @book.route('/delete_review/<int:review_id>')
 @login_required
+@confirm_required
 def delete_review(review_id):
     review_ = Review.query.get_or_404(int(review_id))
-    if current_user.id == review_.author.id:
+    if (current_user.id == review_.author.id and current_user.can(Permissions.DELETE)) or \
+            current_user.can(Permissions.MODERATE_COMS):
         db.session.delete(review_)
         db.session.commit()
         flash("Рецензия удалена", category="warning")
@@ -120,9 +126,10 @@ def delete_review(review_id):
 
 @book.route('/delete_comment/<comment_id>')
 @login_required
+@confirm_required
 def delete_comment(comment_id):
     comment = Comment.query.get_or_404(int(comment_id))
-    if current_user.id == comment.review.author.id:
+    if current_user.id == comment.review.author.id and current_user.can(Permissions.DELETE):
         db.session.delete(comment)
         db.session.commit()
         return redirect(request.referrer)
@@ -131,6 +138,7 @@ def delete_comment(comment_id):
 
 @book.route('/status_up/<int:review_id>')
 @login_required
+@confirm_required
 def status_up(review_id):
     """
     Ставит лайк на пост
@@ -155,6 +163,7 @@ def status_up(review_id):
 
 @book.route('/status_down/<int:review_id>')
 @login_required
+@confirm_required
 def status_down(review_id):
     """
     Ставит дизлайк на пост
